python/momoiro-youtube/requirements.txt
google-api-python-client==2.118.0
boto3==1.34.34
python-dotenv==1.0.0
pytz==2024.1 